+++ src/io/aviso/logging.clj
@@ -4,42 +4,36 @@
   You must [add clojure.tools.logging as an explicit dependency](https://github.com/clojure/tools.logging) of your project."
   {:added "0.1.15"}
   (:require [clojure.tools.logging :as l]
-            [io.aviso.exception :as e]
-            [io.aviso.writer :as writer])
+            [io.aviso.exception :as e])
   (:import [java.lang Thread$UncaughtExceptionHandler]))
+
+(def ^:private eol (System/getProperty "line.separator"))
+
+(defn ^:private install-pretty-logging*
+  [options]
+  (alter-var-root
+    #'l/log*
+    (fn [default-impl]
+      (fn [logger level throwable message]
+        (default-impl logger
+                      level
+                      nil
+                      (if throwable
+                        (str message
+                             eol
+                             (e/format-exception throwable options))
+                        message))))))
 
 (defn install-pretty-logging
   "Modifies clojure.tools.logging to use pretty exception logging."
   ([]
-    (alter-var-root
-      #'l/log*
-      (fn [default-impl]
-        (fn [logger level throwable message]
-          (default-impl logger
-                        level
-                        nil
-                        (if throwable
-                          (str message
-                               writer/eol
-                               (e/format-exception throwable))
-                          message))))))
+   (install-pretty-logging* nil))
   ([frame-filter-fn]
-    (alter-var-root
-      #'l/log*
-      (fn [default-impl]
-        (fn [logger level throwable message]
-          (default-impl logger
-                        level
-                        nil
-                        (if throwable
-                          (str message
-                               writer/eol
-                               (e/format-exception throwable {:filter frame-filter-fn}))
-                          message)))))))
+   (install-pretty-logging* {:filter frame-filter-fn})))
 
 (defn uncaught-exception-handler
   "Creates a reified UncaughtExceptionHandler that uses clojure.tools.logging/error, rather than
-  simplying printing the exception, which is the default behavior."
+  simply printing the exception, which is the default behavior."
   []
   (reify Thread$UncaughtExceptionHandler
     (uncaughtException [_ _ t]