+++ src/io/aviso/repl.clj
@@ -9,8 +9,7 @@
     [clojure.main :as main]
     [clojure.repl :as repl]
     [clojure.stacktrace :as st]
-    [clojure.edn :as edn]
-    [io.aviso.writer :as writer])
+    [clojure.edn :as edn])
   (:import
     (clojure.lang RT)))
 
@@ -29,14 +28,16 @@
   (print-exception e {:frame-limit 0 :properties false}))
 
 (defn uncaught-exception-handler
-  "Creates a reified UncaughtExceptionHandler that prints the formatted exception to `*err*`."
+  "Returns a reified UncaughtExceptionHandler that prints the formatted exception to `*err*`."
   {:added "0.1.18"}
   []
   (reify Thread$UncaughtExceptionHandler
     (uncaughtException [_ _ t]
       (binding [*out* *err*]
-        (printf "Uncaught exception in thread %s:%n" (-> (Thread/currentThread) .getName))
-        (e/write-exception t)))))
+        (printf "Uncaught exception in thread %s:%n%s%n"
+                (-> (Thread/currentThread) .getName)
+                (e/format-exception t))
+        (flush)))))
 
 
 (defn pretty-pst
@@ -120,10 +121,10 @@
   to `*out*` (returning nil)."
   {:added "0.1.32"}
   ([]
-   (e/write-exception* *out*
-                       (-> (copy)
-                           (e/parse-exception nil))
-                       nil))
+   (-> (copy)
+       (e/parse-exception nil)
+       e/write-exception))
   ([text]
-   (writer/into-string
-     e/write-exception* (e/parse-exception text nil) nil)))
+   (-> text
+       (e/parse-exception nil)
+       e/format-exception)))