+++ CHANGES.md
@@ -1,3 +1,8 @@
+## 0.1.37 - UNRELEASED
+
+*Incompatible Change*: Remove the `io.aviso.writer` namespace and change many functions
+to simply write to `*out*` rather than take a writer parameter.
+
 ## 0.1.36 - 22 Dec 2018
 
 Support Clojure 1.10.